admin: add Marco Aurelio to LDAP-only admins (nda)
Per T333884 Marco has signed previous NDAs and is currently
in the process of signing the latest NDA.

He would like to keep his real name out of the public repo
but Legal knows it. So using the "(known to legal)" string that
we have already used before for others.

To be added to the "nda" LDAP group, ultimately for logstash access.

Also a good confirmation for doing T333870 which was requested
by security team.

Bug: T333884
Bug: T333870
Change-Id: I53ba4564c4cf1c6f278d0a86d2225c385b59c11b