install_server: configure root raid only on gitlab-raid1
This change removes configuration of additional SSDs from
gitlab partman recipe. The recipe will configure the first two
SSDs only for the root partition. Configuration of the second raid
for the backup partition will be done later after installation
using puppet.

We were unable to find a working partman recipe to configure
two independent raids across four disks. So we will move the
configuration of the second raid to puppet. This will increase
the complexity of the puppet code but let use use the second pair
of SSDs.

Bug: T330172
Bug: T333674
Change-Id: Ie77560b55076c94bec255f5577e2d75492496d08