[gitlab/ssh] Only recurse if the directory is to be removed
The `recurse => true` was put in place to allow for the directory to be
removed. This, combined with `mode` instead changed the permissions on
the ssh keys recursively to 0755 which is incorrect. Where we have
`$manage_host_keys` set, we set the permissions on the files, but not if
the ssh keys exist already.

This only adds the `recurse` option when the directory is to be
`ensure => absent`, therefore it won't set the permissions on the files
inside the directory by mistake.

Bug: T334736
Change-Id: I530ee2f6fff3715b1dd68f75e1b6c35d7d45bfd5