Place an-worker1132 back into the insetup role
We wish to reimage an-worker1132 and have it assume only the insetup
role temporarily, so that we can rerun the sre.hadoop.init-hadoop-worker
cookbook against it. This is to address an issue with the RAID
controller that we believe has now been fixed.

There is no problem with wiping any data on the 12 data drives, since
this node has been excluded from the active datanode list.

This commit will be reverted once the 12 data drives are correctly
initialized.

Bug: T333091
Change-Id: I5c91d57504cf7a11618125c423210639e849a024