gerrit::migration: ensure gerrit review_site dir exists
In addition to other things also ensure that the
${gerrit_site} (/var/lib/gerrit2/review_site by default)
directory exists before Gerrit has been deployed to a host
the first time.

This allows for rsyncing data when migrating to a new host
before Gerrit has been deployed to it.

But we need to do this only if the directory is not already
defined from the gerrit module, to avoid a duplicate declaration
error on the passive host already using the production role.

Bug: T326368
Change-Id: Id4dbf409708e4833101d57a6dca6ac0e2451a446