Add the perccli utility to the new Ceph servers
The new cephosd100[1-5] servers have been specified with a Dell HBA330
Mini Host Bus Adapter, rather than a RAID controller. Management of this
device is best performed with the perccli utility.

This commit adds the PCI ID of the controller to the pci_ids structure
that determines the content of the raid_mgmt_tools fact on each host. A
new class is added to support perccli_hba controllers such as this.

At present, the class only adds the perccli package to the hosts, but in
due course we will add controller and physical disk health checks to it.

Bug: T330151
Change-Id: I036b60bd2de8cb914669633036a30913a0d82bd5