prometheus: Add support for syncing data between Prometheus hosts
This commit adds a new feature to the Prometheus monitoring system that allows
data to be synced between hosts using the 'quickdatacopy' module. This feature
is particularly useful for performing data migrations, such as during a
distribution upgrade, where it is important to ensure that monitoring data is
preserved.

With this new feature, SREs can easily sync monitoring data between hosts,
ensuring that data is not lost and that monitoring remains uninterrupted.

Bug: T309979
Change-Id: Ia72f85a736ee6be290516e67022b3a0b92f9503e