gitlab: add script to create fs and raid for backup partition
This change adds a script to configure the two additional SSDs
to be used as backup partition for /srv/gitlab-backup. As mentioned in
Ie77560b55076c94bec255f5577e2d75492496d08 we were unable to implement
this in partman, so bootstrapping the backup partition was moved to
a dedicated script.

Bug: T330172
Bug: T333674
Change-Id: I8766182572f2a3da2916b85c7ec3226b60ba5119