mariadb::generic_server: change default datadir path
In the past there has been considerable extra work
and surprises for direct or indirect users of this class.

That is because it defaults to /srv/sqldata and not /var/lib/mysql
which is the default path picked by the distro package it installs.

And the class does not ensure the service gets restarted.

So this means people would start using it, everything seemed fine
at first but at some later point when the service or VM gets
restarted, suddenly the datadir is different and databases appear
to have vanished.

After having fixed this for all users of the "simplelamp2" role
which used this class and all other remaining users, we can
as the final action flip the default here without affecting anyone.


This has nothing to do with production db servers, it has only a
few users like phabricator, phorge and a handful of projects in cloud VPS.


Bug: T329571
Change-Id: I8dcdde07402781e1857bf4d82d9d432567b62643