cloudgw: allow VMs to speak to new gerrit server (gerrit1003)
This is adding the IP of the upcoming replacement gerrit server
so that cloud VMs can talk to it just like they can talk to the
current production Gerrit server.

Right now the DNS name is "gerrit-new.wikimedia.org" but it will
become gerrit.wikimedia.org when we switch so not bothering with
the comments here.

Adding the IP now. Removing the old IP will be part of decom'ing
gerrit1001 so it's not a straight replacement.

Bug: T326368
Change-Id: I3c01b398e632640e0c31b2f20fbe2b99b466ece4