git-sync-upstream: add support for gituser and alternate base directories
This Cr adds parameters so that the git repos can be installed into a
different location namely /srv/git which will be used by puppetserver.

We also add a parameter to specify the git user, the systemd timer runs
as root (which is currently needed to update the prometheus file)
however the git repos are owned by gitpuppt.  This causes issues with
the most recent version of git.

The current WMCS puppetmasteres work around this later issue by setting
safe.directory[1] however it would be better to move to a more robust fix

[1]profile::openstack::base::puppetmaster::safe_dirs

Change-Id: Id9e29b358441e02af6ef6b1f0f0947ffb999cb7c