admin: move fnavas to ldap_only admins, remove from a-privatedata-users
technically reverts I3ed279a1614ff but moves the user to another section

After a lengthy ticket we figured out by comparison to another
existing user that this was not actually needed and by itself
did not give the access desired.

What is actually needed is the 'wmf' LDAP group and that by itself
should do it.

Ccompare to existing ldap_only user sdelbecque, who have only that
but can see what this user wants to see as well.

So, moving from "shell" section (in quotes because it's the
"analytics-privatedata-users without keys"-thing) to ldap_only
section.

No fields for UID, key etc needed. Just email address.

We are keeping the expiry_date and contact as it is.

Bug: T331482
Change-Id: Id3a471f3555e84ec5d74c417d3619ec2da254504