Add a custom ceph_disks fact
In order to deploy the CEPH OSD daemons on the new servers we need to be
certain which devices are being referenced during each operation. We
cannot depend on the standard /dev/sd* device names, since these may
change across reboots.

The ideal mechanism for doing so is to use the /dev/disk/by-id/wwn-*
which is a symlink to the correct device, based on the WWN value of each
drive.

However, it is not possible to ascertain simply from the WWN values
whether these are HDDs, SSDs, or whether they are the drives that are
used for the O/S mirror.

This custom fact aims to resolve this problem by creating a new custom
fact on the cephosd servers, containing data on populated storage slots
along with the wwn, serial number, medium, and interface of each.

This fact will then be used by the ceph module to provision OSDs
automatically. It is deployed to all hosts as part of the wmflib module
because it is specific to our hardware choices, but the fact will one be
created on servers where the perccli64 binary is available and the
ceph-osd package is installed.

Bug: T330151
Change-Id: Ie022ed908edc37d78649fb19d5fb44606c60f93c