httpd: always use systemd
We should just use systemd by default.  This is the default in puppet and
debian for everything we use.  Further the sysv sc4ripts call systemd so
this just adds an un-eccesary layer of indirection

This was also discussed in:
 Ie56678c95eccacc3b0255926a2ba7a58459dabf4
 Ice9d26df6226c8b36e391eddd4b75f40982f3bdd

Bug: T331706
Hosts: O:mediawiki::canary_appserver
Hosts: O:mediawiki::appserver
Hosts: O:lists
Change-Id: I851f19a9a5c3d107e667d408053969e21cdd85d4