Varnish/ATS semicolon workaround for Restbase
This is the same workaround we applied in a26649693 for MediaWiki
traffic in response to T238285 , now also applied to Restbase
traffic.

This is a workaround rather than a proper fix, and it's happening
at the wrong layer (Varnish, which isn't really causative here),
but it's what we can most-easily do for now.

The /real/ issue is a confluence of two separate factors that
shove us into a corner case:

1) That MediaWiki canonicalizes semicolons as un-encoded in titles
   and also allows them to be the trailing character of titles,
   both of which are legitimate choices according to current
   standards, but may be relatively-rare on the Internet in
   practice.  All of our other caches and revproxies handle this
   fine, other than ATS.

2) ATS's URI parsing is of archaic origin, and settled long ago on
   an interpretation of this case rooted in some older RFC, and is
   likely stuck with it as a default, at least.  When it sees a
   bare un-encoded semicolon at the end of the Path section of a
   URL, it treats it as a param delimiter with no following param,
   and then deletes the trailing semicolon delimiter.

Eventually, we should probably pursue fixing this in upstream ATS,
by having some non-default option to interpret the trailing path
semicolons like other modern revproxies, but it may be tricky.

Bug: T238285
Bug: T334467
Change-Id: Ib4afb623ca76e235c31c929c364a29a30167a2db