prometheus: Add label to prometheus6001 data blocks to prevent data duplication
This commit adds a label to the prometheus6001 data blocks to allow the instance to act as replica
in the datacenter. This will prevent Thanos from indexing duplicate data blocks that contain the
same data but have different IDs, which can occur when multiple Prometheus instances scrape the same
targets.

Bug: T335406
Change-Id: I0f8b9a288bd5d2ba244e98a8f20ab3c7a6ecbd7e