gitlab: enable and run partial backups daily
This change enables partial backups on GitLab hosts. As described in
T316935#8805128 "packages" take up most of the backup space. So this
change enables partial backups, excluding packages from the backup.
The goal is to do more frequent repository and database backups without
the overhead of hour-long backups.
The name of partial backups is fixed to partial_gitlab_backup.tar using
the BACKUP environment variable. This also helps to exclude this backup
from bacula filesets.

Bug: T316935
Change-Id: If79a737b016effdf1037eaa81b5cf13e06a8120c