mediabackups: Set both dcs to backup Commons by default
This has to be changed so it includes one (or several) backup
source instance for each section to be backed up, but for now
setup a valid backup source & Commons wiki as the default wiki
to backup files from.

Also reduce the batch size to 100, now that almost all files
are backed up, speed up is not that important, and a smaller
batch prevents mysql query timeouts.

Bug: T327157
Change-Id: Ic3dfe9c8b34366b580a71cfe89cf97f44b81a903