pybal: Fix hostnames not being sent on alert
If the IPVS pybal checks found differing IPs, alerts would fire; it's
very unlikely that hostnames would differ but not IPs, so the IP alert
would be triggered first and then promptly exit, not ever sending
hostnames.

To accomodate alerting/highlighting on IRC (and just general clarity),
this will alert with hostnames in addition to the IPs.

Bug: T322377
Change-Id: I88e4ed8ca3e7d2ce2b53701b5d0961b8917f43c7