nfs-exportd: Don't crash out if a dns lookup fails
This script iterates through a list of nfs servers to determine
which one is us. Previously if there was an invalid entry in the
list we crashed on the dns lookup, which might've prevented exports
from being updated properly. Now, log and continue.

This patch also changes a couple of log messages from warning
to debug because I briefly thought they were important when
they're actually normal behavior.

Bug: T335336
Change-Id: Iaa1e556873b962f305f6ac2e0809f2eb38d1ef8c