gerrit: accept http status 404 in blackbox http monitor, for now
The gerrit-replica.wikimedia.org host can be expected to
return a 404, as opposed to a 302/200.

For now this is the normal status quo.

Unless we want to not monitor the replica or change that
it does the 404, we have to accept it in monitoring.

For now just avoid that the freshly added monitoring alerts
people.

Bug: T329587
Bug: T331901
Change-Id: If5fe2fff7d3cb9b01bc27b5ea1b3a8284abfe949