gerrit: do not monitor the replica
The gerrit-replica host behaves too different from the
regular hosts to monitor them in the same way.

Either we don't monitor the replica at all, or we make it
NOT return a 404 and "Not found".

Or we have to accept 404 and "Not found" as normal even on
the regular hosts.

Probably the fix is to have separate check for the replica host
but also make it return something useful on the main page.

Bug: T329587
Bug: T331901
Change-Id: I85ff852e7d83255e02187d38c9850407c175fd84