search: Fix collection of *_titlesuggest metric on small clusters
This metric was only being collectd on the large clusters. We only had
per-index stats enabled on those clusters and the *_titlesuggest data
collection was mistakenly behind a check that there were indices to
collect per-index stats from.

Adjust the handling so *_titlesuggest is always collected, even when
per-index stats are not enabled.

Bug: T327199
Change-Id: I5b05e0cb43ace3b8da7fd2ff46362cf6538db421