kubeadm: install certificates before trying to use them
Previous to this patch, there was a dependency issue with etcd-server
needing the certs but the certs requiring the package.

I guess the dependency on the package was there for the directory tree,
which is created by puppet a few lines above in the same manifest.

Simply install the certificates before trying to use them.

Bug: T335052
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: If9216a235b814e2efd737d84b8685fba1977cebb