prometheus: Synchronize only the /srv/prometheus directory when migrating data
The change was made to avoid ownership conflicts that were occurring during synchronization due to
different UID/GID mappings of instances. The other folders in the /srv directory are created when
the instance role is applied, so syncing them causes issues. By syncing only the /srv/prometheus
folder, we ensure that the Prometheus data is up-to-date without causing any ownership conflicts.

Bug: T309979
Change-Id: I287090199b09cca8b2251d7195193b0bbaada3ca