profile::bird::anycast: allow setting the BGP IP address from the profile
Previous to this patch, the first IP address of the host was being used to
configure the bird anycast.

In some servers (for example, cloudlb) this is undesirable, as they need to have
a different IP address set there.

This patch adds a new 'ipv4_src' parameter in the bird profile that is later
passed to the bird base class. This allows to override the default in the
bird base class.

Existing clients of this code should see this as a NOOP.

Hosts: auto
Bug: T335760
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I3650ec5ef5df76aef644ace5056e5d9312b745f3