gerrit: disable replication from gerrit1003
There is no more general "active host" setting that defines
a single machine as the one that should replicate.

But there is code that only adds the remotes to the config
if $replication is set. So this will disable it on gerrit1003.

The list of replication hosts may not be used anymore.

We should probably disable it by default and only enable it explicitly
in a separate step.

Bug: T335730
Change-Id: I4844c01d3e6f59299a56fa896e1cf83c8ebad26b