ci: use an array to manage gitcache repos
Adding a new repository to `profile::ci::gitcache` involves a large copy
paste when it should be solely adding an item to a list.

Remove the obsoletes `require` parameters, they date from when we had
Puppet executing resources in the order of the title checksum, we
nowadays use the manifest order and thus the parent directories are
created first.

An example adding two repositories is in child change
I5f5687cebeaf41c1f0e1038b0567c48e6c19a05c

Change-Id: I2ebac179f7cdc9321040dc6902efb09831e16d9c