cloudlb: introduce haproxy check for the BGP VIP
Instead of checking icmp (which is an effetive always-true) use
the haproxy check in alive mode.

The alive mode will leave the BGP VIP running as long as HAproxy is running.
There is another mode --check=failover that will detect if a single backend
is down and remove the BGP VIP.

We can switch to that mode later on.

Bug: T324992
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ie57780312b21e3baa6c34811589c5cd0758020b2