haproxy: check_haproxy: don't use a tempfile
The haproxy socket output is fairly small, it can fit into a variable, removing the
need for any disk operation. Moreover, before this patch the code was using a fixed
path for the tempfile, which is insecure.

While at it, modify the output from:

 user@server:~ $ /usr/local/lib/nagios/plugins/check_haproxy --check=failover
 CRITICAL check_failover servers up 0 down 63:\nbarbican-api_backend:cloudcontrol2001-dev.wikim[..]

To:

 user@server:~ $ /usr/local/lib/nagios/plugins/check_haproxy --check=failover
 CRITICAL check_failover servers up 0 down 63:
 barbican-api_backend,cloudcontrol2001-dev.wikimedia.org
 barbican-api_backend,cloudcontrol2004-dev.wikimedia.org
 barbican-api_backend,cloudcontrol2005-dev.wikimedia.org
 barbican-api_backend,BACKEND

(i.e, one DOWN service per line, which is more readable)

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I4bdb1ce1e83977061b98131d2c5cd619fabd30fa