trafficserver: make mw-on-k8s use a config file
Much like we did for multi-dc, make the mw-on-k8s module read a
configuration file at runtime so that we avoid having to reload trafficserver
every time we want to change something there.

Also, refactor tests so that we can bypass loading the configuration file
like we did in multi-dc.lua.

Finally, bump tlslua to use python3 so that it can work on modern computers.

Bug: T336037
Change-Id: I3d51952aa005a4b1a06a5fe813610760f3145dd8