varnish: Set Cache-Control: no-transform header
When the user access Wikipedia through Chrome Lite, pages will kick in
if the page takes too long to load. When that's the case, responses will
not be stored in the browser cache. That means that the next page that
the user visits will be a cold cache again and will take longer time to
render, causing the Lite page to kick in again. This will happen all
over again. The normal behavior for users accessing Wikipedia on a slow
connection is that the first page is slow, but then we have stored
JavaScript and CSS in the browser that makes the next page view faster.

By setting this header we can instruct Google (and other similar
products) to avoid such behavior.

https://tools.ietf.org/html/rfc7234#section-5.2.1.6

Bug: T218618
Change-Id: I6b1fbdfb381fbcebc53b494c3040f121f7ea2167