Add two ldap_only users from bishopfox
These two user accounts are owned by named security analysts working
for Bishop Fox, who are working with us on an application security
assessment.

They will temporarily be granted access to the `nda` group in LDAP, so
that they may access the tools they require as part of the engagement.

Bug: T336357
Change-Id: Ib222abe23d0eda3ee26568625b0c7f4441187e99