gerrit: enable replication from gerrit1003, disable from gerrit1001
As part of the migration of gerrit1001 to gerrit1003 we have
to switch which becomes the source of replication.

Disabling it on gerrit1003 was the fix for recent bug T335730
and we need to avoid that both run at the same time.

So this needs to flip during migration window after rsync.

Bug: T326368
Bug: T335730
Change-Id: I145b8e08703c971424678cd3df485b3d6e05db89