Add Airflow configuration to connect to DataHub
We are emitting to Kafka by using the schema registry, Karapace.
Then the metadata are fetched by the mce-consumer service on
the DataHub side.

The core of this patch was already included in: 931683

Now, remains a configuration on all Airflow instances (except
analytics_test) which avoids the automatic emission of metadata
about the dag execution to DataHub. Whithout this configuration
you currently see an error at the bottom of each task
execution log:

        airflow.exceptions.AirflowNotFoundException:
        The conn_id `datahub_rest_default` isn't defined

Note for deploy: the new configuration added to airflow.cfg file is
goint to be picked by the scheduler without having to restart
the process.

Bug: T333004
Change-Id: I9119f2816747795a17d065c588affbc559102da8