prometheus/k8s: add selective scraping of ports in staging
We want to be able to signal clearly to prometheus which ports to scrape
in a pod, via a naming convention.

To this end, add a configuration for scraping that includes any port
labeled `-metrics` *and* has a specific annotation set
to true.

This allows us to create a seamless transition for the services that we
already collect metrics for but we want to transition to the new system
- just substitute:

prometheus.io/scrape: true

with

prometheus.io/scrape_by_name: true

Please note, the new scraper is for now only applied to k8s-staging
in order to test the setup.

Bug: T271822
Bug: T318707
Change-Id: I9e747a0202e0dce5c5ac64eacd85c92a9353e0bf