benthos: use kafka_franz for the webrequest_live instance
In https://github.com/benthosdev/benthos/issues/1806 the upstream
devs added support for batching and select topic partitions to
kafka_franz, a new component that seems preferred from the kafka
one. I tested the new Benthos version 4.15.0 on stat1004 and
it worked nicely.

The idea is to replace the kafka component/client with kafka_franz,
see how it goes and then reduce the topic partitions (and adjust
sampling) later on.

Bug: T331801
Change-Id: Ib2c05e3dcd0632ff512ca382e269e8b6a720b591