Remove python-is-python3 package from hadoop bullseye hosts
This workaround was added in order to try to prevent hive from depending
on the python2 packages on bullseye, but it didn't work as required.

What happened was that on alternate puppet runs, python-is-python2 and
python-is-python3 were being installed, resulting in errors from hive.

A different change has allowed us to keep python2 available. See:
https://gerrit.wikimedia.org/r/c/operations/puppet/+/901595

Therefore, this failed workaround can be removed.

Hosts: an-test-worker1001.eqiad.wmnet,an-worker1078.eqiad.wmnet
Bug: T336281
Bug: T329363
Change-Id: I23ea61685c34a2f9a562a176dd39877c942e1c42