gerrit: allow masking the service and do so on gerrit1001
This will allow us to re-enable puppet on gerrit1001 while
being sure the gerrit service never starts.

Not even accidentally, so we mask it.

The way this is done via additional service_params ensure and enable
plus the "systemd::mask/umask" class is already used in other places.

Compiler output shows the diff between 1001 and 1003/2002 is as intended.

Not using any default value on purpose per style guide.. I guess.

Bug: T326368
Bug: T336427
Change-Id: I27c5a64909ddcd1b220c09eb1a1ff3f356f29b7c