gerrit: add gerrit1003 SSH host key known_hosts
Looks like we should add gerrit1003 here.

Though.. it probably just matters for the hosts
that are replication destination, not sources.

This would also explain why the one for gerrit1001 doesn't match
what's in the repo but it doesn't seem to have bothered us.

For gerrit2002 the one in the repo matches what is in
/etc/ssh/ssh_hos_rsa_key.pub on the machine, so doing there here
as well.

This is about openssh on port 22, not MINA ssh on 29418, as far
as I see, for inter-gerrit ssh communication.

So I guess it also doesn't really matter unless we ever want to
replicate TO gerrit1003, like in a DC switchover that Gerrit would
take part in.

Bug: T326368
Change-Id: I9920b08a934b9749f023449296380d1222021355