site: remove gerrit1001 from gerrit role, rm hiera host data
Move the server back to the "insetup" role at the end of
its lifecycle.

This will remove gerrit specific configs and also
shell access for gerrit admins.

Data from /srv/gerrit and /var/lib/gerrit2/ has been copied
to gerrit1003 and added to Bacula as well.

All the hiera data is cleaned up too.

But it does NOT yet kill the file system or shut the server
down.

Bug: T336427
Change-Id: I142b87ea128946f7c2276c6f6c3cc3d930e7afff