Update the version of airflow on an-test-client1001
We are testing a new version of airflow and wish to test this on
an-test-client1001. The package isn't currently uploaded to
apt.wikimedia.org but it will be installed manually.

This override of the package version stops puppet from trying to
downgrade the package to that available on the repository.

Bug: T336286
Change-Id: Ie9106f869b2d539227ca1b4c5da9e6de776eac94