gerrit: remove lfs_dir parameter, use hardcoded new default
We switched the path to LFS data to move it out of the root file
system for T333143.

old path: /srv/gerrit/plugins/lfs
new path: /srv/gerrit/data/lfs

The new path was first configured for gerrit1003 which is now
the production server.

gerrit1001 is not production anymore.

And as we established in today's meeting, LFS data on the replica
host is irrelevant.

Finally the cloud instance for gerrit is currently down for other
reasons and has to be redone or brought back regardless.

So we can now make the new path the default and clean up any
hostname-specific "hacks" to complete this process.

Bug: T334521
Bug: T333143
Change-Id: I1af168d4850ca0ee20e8ae8e142ea214f2935848