httpbb: add simple tests for gitlab.wikimedia.org
These are 3 simple tests for httpbb to check if gitlab.wikimedia.org
is up and working.

a) check that / is a 302 and redirects to https://gitlab.wikimedia.org/explore

b) check that /explore/groups is a 200 and contains the string "repos"

c) check that /help is a 200 and contains the string "GitLab"

Usage example:

[deploy1002:~] $ httpbb --hosts gitlab.wikimedia.org ./test_gitlab.yaml
Sending to gitlab.wikimedia.org...
PASS: 3 requests sent to gitlab.wikimedia.org. All assertions passed.

If you can think of other things to test to add to this or replace
something let me know.

side note: for other services we can run these tests against individual
host names like "check if host1234 returns the right things for virtual
host foo.wikimedia.org". But you can't currently do that with gitlab
servers because of how well they are locked down by firewalls.

Bug: T326891
Change-Id: I6f53603ad49c3a647e10fc4ad7f0e36c72195a99