mwopenstackclients: use novaobserver creds to determine domain of a project
We have a chicken/egg issue here where we need to know what domain
a project is in before getting a token, but we need a token
in order to find out the domain. With this patch, we always
use the novaobserver creds to determine domain, so that we
can still get clients for creds that can't otherwise query
keystone about domains.

Change-Id: I3ce01c19fd453d2b878f4b7035bef07ccc651506