Add an extra property 'CollectMode' to each user's jupyter service
We wish to reduce the amount of log noise associated with individual
users' jupyterhub-singleuser notebook servers.

In order to do this we set the CollectMode to 'inactive-or-failed'
instead of the default 'inactive'.

This will cause systemd to forget about the unit as soon as it has
exited, even if exited abnormally.

Bug: T336951
Change-Id: Ibc165561a6f3ae7269e77d12bf3212a5b2c03914