cloudgw: refactor interfaces setting to use the base module
This feels a bit more elegant. Instead of coding them in the template file.

The movement is possible because now interface::post_up_command exists,
which wasn't the case when this setup was first introduced.

While at it, drop IPv6 forwarding which is not required at this point.

Bug: T347687
Bug: T338334
Hosts: re:cloudgw.*
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Icf3fbdca984c2d74009ce405ee1bc78b8fe89eef