profile::kafka::mirror: add support for PKI certificate
Kafka mirror runs as a daemon on various kafka nodes, and
it authenticates to each broker with a cergen TLS certificate
carrying the CN: kafka_mirror_maker, that we use as username
to match various Kafka ACLs.

This change adds the possibility to use PKI to ease the process
of TLS certificate renewal.

Bug: T337248
Change-Id: I73640add5594e95825b9b0642f69546039002171