microsites: remove http blackbox monitor for 15.wikipedia.org
This alert should not be associated with the microsites profile
anymore.

It is currently alerting because 15.wikipedia.org moved to k8s
and was removed from these previous backends we monitor here
with I898a80e6953d2f.

I wonder though if this means a step back since we just worked
to have this monitoring for every site and now would anything
replace it?

Bug: T300171
Bug: T337424
Change-Id: I7c533a4308a84088a54911dd1ddfb913395766b0