spicerack: add test-cookbook script
Why:
In order to be able to test a cookbook's Gerrit patch, SREs are
currently cloning the change on the cumin hosts and then running it with
some custom-tuned configuration file for the cookbook binary.
This patch provides a canonical way of testing this kind of patches with
auditing logs and !log to SAL.
The initial draft was made during the Wikimedia Hackathon last week.

What:
* Checkout the Gerrit change to the user's home in a deterministic way.
* Allow to specify a specific patch set instead of the last one.
* Allow to delete a previously checked out change
* Automatically detects local modifications and allow the user to choose
  to run with the modifications or abort.
* Store the logs in the user's home but in a predictable path that is
  easy to grep in case of auditing.

Change-Id: I6e0d3d4a5ebf7dfc0d69a761d13624f82199319a