varnishkafka: add catch all systemd unit
The varnishkafka instances may need to be all restarted in the future,
for example if a new TLS certificate is created by PKI because
the current one is almost expired. Instead of manually roll restarting,
we can use a catch-all systemd unit to use in puppet notify
to trigger a restart of all units.

Bug: T337825
Change-Id: Id764484e441a7b392ba57868960063d4b575d0ae