profile::cache::kafka: add support for PKI
The current TLS cert used by all varnishkafka instances is a cergen
one, that has a special CN used by Kafka broker in their ACLs as
username to establish if a producer can or cannot send messages
to a topic. This change introduces the possibility to use PKI
to generate the TLS certificate, adding also an automatic restart
of all the instances when the cert is issued.

Bug: T337825
Change-Id: Id28b15320c134f82e894c0397f9f464bfb5976f7