wikireplicas: restore pybal monitoring
Merge after all the maintenances in T337446 are complete, to
restore the same healtchecking it had before.  This is a partial
revert of Id5f8033f509c5db19b65f82b6727b80c3775453f - it reverts
just the wikireplicas-specific part, but leaves alone the general
capability to have a monitor-less service.

Bug: T337446
Change-Id: Idec791932a0ae533a9224eabec559373e7a571c8