gerrit/bacula: adjust Gerrit file paths to be backed up
- /srv/gerrit/plugins is empty now, lfs data has moved to ./data/
- /srv/gerit/data is 20G
- /srv/gerrit/git is 46G
- /srv/gerrit total is 215G - it contains multiple older copies of
  the git repos (see T336427#8890714) but those are not backed up

- /var/lib/gerrit2 is copied during migrations from one host to
  another but not backed up in Bacula.

Therefore:

- back up the entire /srv/gerrit and everything inside it, remove
  old backups whenever seems appropriate

- back up /var/lib/gerrit2 too so we feel like we can recreate a
  Gerrit server from scratch the same way we would do a migration.

If there are concerns about the increased size, consider that this is
just supposed to last until we decide the multiple old git copies
are not needed anymore.

Regardless of this being in Bacula, it already has been copied from
gerrit1001 to gerrit1003, so in no case should the decom of gerrit1001
be blocked.

Bug: T336427
Bug: T326368
Change-Id: I674fe8a5e2a6b73738ff853b11bf929d99cbc026