zuul: add a gerrit-reporter gerrit connection
The Zuul scheduler has two connections to Gerrit:
- a source one which is the stream events interface to receive events.
- a reporting one to do reviews of changes

Our configuration has both kind under the same name `gerrit` and the
connection instances are stored in a dict which is used to retrieve the
instance:

    {
      "gerrit": <GerritConnection #1>
    }

When the source fails and reestablishes the connection and at the same
time a reporter tries to emit a report, they both receive the same
instance (albeit in different threads) and enter a race condition which
causes a variable to be replaced while the ssh connection stays
established. As a result we have an extra useless connection idling
which goes against the 4 ssh session per user imposed by our Gerrit and
leads to zuul merger not being able to establish a connection.

To workaround the lack of lock on the GerritConnection variable, this
creates a new connection `gerrit-reporter` on the Zuul server.
Internally it would thus be stored under a different key:

    {
      "gerrit": <GerritConnection #1>,
      "gerrit-reporter": <GerritConnection #2>
    }

The source and reporter would thus no more share the same instance and
that should workaround the race condition.

This should not affect any operation until we change the Zuul layout in
integration/config to start reporting using this new connection:

      pipelines:
        - name: test
          success:
    -       gerrit:
    +       gerrit-reporter:
              verified: 2
              tag: autogenerated:ci-test

It should be a noop for the zuul-merge.conf.

Bug: T309376
Hosts: contint2001.wikimedia.org
Hosts: contint1002.wikimedia.org
Change-Id: I0e7369b8f6c533498c2abad75ab7da439a94572e