codfw1dev: remove traces of cloudcontrol2004-dev.wikimedia.org
The server is the last in the migration to the cloudlb setup.
As part of the change, is losing the wikimedia.org domain in favor
of .codfw.wmnet & private.codfw.wikimedia.cloud.

The profile::openstack::codfw1dev::haproxy manifest is no longer in use,
so drop it. It has been replaced by the already-in-place cloudlb role.

Bug: T337828
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ie8bc319dcd67688018968420567705b1445e40a7