k8s::kubeconfig: Don't specify certificate-authority
With certificate-authority set, python3-kubernetes clients (like
imagecatalog and cookbooks) are no longer able to validate the
certificate of the kubernetes api servers.
My guess is that the code does only use that exact file for validation
and PKI does not include the complete chain but only the intermediate
(even in the "chained" file).

As we ship the chains in wmf-certificate package anyway should be safe
to remove the certificate-authority from kubeconfig.

Follow-up of: Id70da3328eda2b3e5cd27f80ab2a54130bd8c399

Hosts: R:k8s::kubeconfig
Bug: T325268
Change-Id: Ibcf08dc30e8518e385335048801304ac3d3e6b3e