Remove runphpscriptletonallwikis.py
This script was introduced as a helper for listwikiuploaddirs.py, which
was removed in 8016f3a23a4a27e1. The other relevant scripts were
removed, and installation of the script was removed, but deletion of
the script was missed.

I'm thinking about adding a parameter to getMediaWikiCli(), and this
script is causing me some head scratching.

Change-Id: Ib5d8e40a638a71af2db3e3271418d321e0ba58c5