Extend KDC logrotate config to also include krb5kdc.log
With the expansion of the Presto cluster, the KDC hosts get a lot more
TGQ_REQ requests and following the update to Bullseye we're getting an
increased rate of deprecation warnings (which will need a refresh of
Kerberos principals).

Update the KDC logrotate config to also include krb5kdc.log in the
daily rotation/compression.

Hosts: krb1001.eqiad.wmnet
Bug: T337906
Bug: T337544
Change-Id: Iab3f0f75a21c3112301297dad06bd1025afba638