expand_path, regex_data: use yaml safe_load when available
Ruby 3.1 and psych 4.0 default to using safe_load even when load is
called, this causes an exception for expand_path and regex_data. Switch
to safe load on rubies greater than 2.7. This does not affect puppet
server 7 as it uses a different java yaml implementation, but this
does affect puppet applies on Debian bookworm.

Bug: T337972
Hosts: P:base::production
Change-Id: I53790ff92ba177e5c41a3f59a3d62464cfd5285d