nove vendordata: purge systemd-resolved libnss-resolve during first boot
These packages seem to be included in the default upstream
Bookworm packages. They behave differently from the resolvers
we're familiar with, and systemd-resolved isn't recommended to
be installed.

The main issue with systemd-resolved is its refusal to resolve
standalone hostnames, specifically 'puppet'.

This additional command should be harmless on Buster and Bullseye
which don't have the packages anyway.

Bug: T338192
Change-Id: Ia8d2a4cb1adf08190317a7e0cf6c8c8c3d58a28a