fix-staging-perms: Change group owner to deployment
scap stage-train runs as mwpresync, which is a system account
assigned to the mwbuilder and deployment groups. The scap command
writes to /srv/patches and /srv/mediawiki-staging to prepare the
train. However, with the directory ownership set to wikidev, it is
unable to do so, as wikidev is for human accounts.

All deployers should be in the deployment group, so let's use
that instead.

Bug: T338205
Bug: T338180
Change-Id: Ie5f5919ed2281424bc9e51ce16a35fe0b7e0151e